refactor wanted, add QR image section
