+++ src/pages/_document.tsx
@@ -14,7 +14,8 @@ class MyDocument extends Document {
             href="https://fonts.googleapis.com/css2?family=Nunito:wght@500&display=swap"
             rel="stylesheet"
           />
-          {/* <link
+          {/* slick carousel */}
+          <link
             rel="stylesheet"
             type="text/css"
             charSet="UTF-8"
@@ -24,7 +25,7 @@ class MyDocument extends Document {
             rel="stylesheet"
             type="text/css"
             href="https://cdnjs.cloudflare.com/ajax/libs/slick-carousel/1.6.0/slick-theme.min.css"
-          /> */}
+          />
         </Head>
         <body>
           <Main />
